changing to long

+++ src/main/kotlin/br/com/acmattos/hdc/scheduler/port/rest/ScheduleCommandController.kt
@@ -96,7 +96,7 @@ data class PeriodRequest(
     val weekDay: String,
     val from: String,
     val to: String,
-    val slot: Int
+    val slot: Long
 ): Request<Period>() {
     override fun toType(
         who: String,
